Add header styles
I added the universal selector to remove the default margin and padding. Also added border-box property to make containers intuitive.

The strategy for writing the CSS has been like this: begin from the very beginning of the DOM structure. That means address the additional margins and paddings from the outset. Also, handle the body first.

When generic things are taken care of, just focus on the first element to style. I focused on header for this sprint. I also demarcated the beginning of the header styles by comment. This would allow me to organize and identify duplicates later on. At least, that's the idea.