Complete styling container one
Apart from a few changes in the html, class-wise, there was no other change in it. I focused on styling container one for this sprint.

Adding styles to the button was interesting work. Just typing button in the declaration also provided most of the options that I needed. So I could follow-along by referencing the reference image to create the styles.

Once all containers are styled, I will work on the copy.