Update web copy and minute style changes
Save point: if you want to see the version before writing copy, pls refer to the previous commit.

Writing a full copy was a good idea. Because some of the elements that I thought were working just fine broke down as soon as I added multiple lines of content.

But mostly, the problem was related to not using max-width when it comes to text chunks. I also used <br> tag to break important sentences that needed to be on a new line for maximum copy-impact.